# src/main/resources/specs/collector-manager-api-spec.yaml
openapi: 3.0.3
info:
  title: Recolnat
  description: |-
    Le réseau national des collections naturalistes 
    est une infrastructure de recherche française dont l'objectif est 
    de produire un corpus de données basé sur les collections d'histoire naturelle de France, 
    permettant de valoriser les recherches au service de l'étude de la géo- et de la biodiversité actuelle et passée.
    Liens utiles:
    - [RECOLNAT](https://fr.wikipedia.org/wiki/RECOLNAT)
  version: "1.0"
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@mnhn.fr
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
security:
  - collection-oidc: [ ] # <--- Use the same name as specified in securitySchemes*, here without scope
paths:
...
  /v1/institutions:
    post:
      operationId: addInstitution
      tags:
        - Institution
      summary: Institution integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstitutionRequest"
      responses:
        '201':
          description: Successful operation
          headers:
            institutionId:
              description: Institution Identifier created
              required: true
              schema:
                type: string
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    get:
      operationId: getInstitutions
      parameters:
        - name: page
          required: true
          in: query
          schema:
            type: integer
            format: int32
            minimum: 0
            description: number of page to retrieve
          example:
            id: 1
        - name: size
          required: true
          in: query
          schema:
            type: integer
            format: int32
            description: number of elements per page to retrieve
          example:
            id: 10
        - name: q
          required: false
          in: query
          description: String to look for in institution name
          schema:
            type: string
        - in: query
          required: false
          name: partnerType
          schema:
            enum:
              - PARTNER
              - DATA_PROVIDER
              - MEMBER
          description: partnerType filter (PARTNER, DATA_PROVIDER, MEMBER)
        - name: columnSort
          required: false
          in: query
          description: Champ de tri
          schema:
            type: string
        - name: typeSort
          in: query
          description: Ordre de tri
          schema:
            type: string
            enum:
              - ASC
              - DESC

      tags:
        - Institution
      summary: Retrieve institutions
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionDashboardResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/institutions/{institutionId}:
    get:
      operationId: getInstitutionDetails
      tags:
        - Institution
      summary: Retrieve collections of intitution
      parameters:
        - name: institutionId
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: institution id
          example: 386ff0b0-a6cc-11ec-b909-0242ac120002
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionDetailResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    put:
      operationId: updateInstitution
      tags:
        - Institution
      summary: Update intitution
      parameters:
        - name: institutionId
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: institution id
          example: 386ff0b0-a6cc-11ec-b909-0242ac120002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstitutionRequest"
      responses:
        '200':
          description: Successful operation
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/institutions/{institutionId}/logo:
    patch:
      operationId: addLogoInstitution
      tags:
        - Institution
      summary: Add Logo Of Institution
      parameters:
        - name: institutionId
          required: true
          in: path
          schema:
            type: string
            format: uuid
            example: 386ff0b0-a6cc-11ec-b909-0242ac120002
      requestBody:
        required: true
        content:
          multipart/form-data: # Media type
            schema: # Request payload
              type: object
              properties:
                logo:
                  description: logo file to be uploaded
                  type: string
                  format: binary
              required:
                - logo
      responses:
        '201':
          description: Successful operation
          headers:
            institutionId:
              description: Institution Identifier created
              required: true
              schema:
                type: string
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/institutions/partners:
    get:
      operationId: getInstitutionPartner
      tags:
        - Institution
      summary: Retrieve all institution partner
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerResponse"
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/institutions/options:
    get:
      operationId: getInstitutionOptions
      tags:
        - Institution
      summary: Retrieve institutions
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InstitutionOption'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
...
components:
  securitySchemes:
    collection-oidc: # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: openIdConnect
      openIdConnectUrl: http://sample.com/realms/recolnat-oauth2/.well-known/openid-configuration
  schemas:
...
    InstitutionDashboardResponse:
      description: Institution Response
      type: object
      nullable: false
      properties:
        totalPages:
          type: string
          format: int
        numberOfElements:
          type: string
          format: long
        data:
          items:
            $ref: "#/components/schemas/InstitutionDashboard"
...
    InstitutionDetailResponse:
      description: Institution Response
      type: object
      nullable: false
      properties:
        id:
          type: string
          format: uuid
          description: id sous forme UUID de l institution
          example: 50f4978a-da62-4fde-8f38-5003bd43ff64
        code:
          type: string
          nullable: false
          description: code , libelle court de l institution
          example: MNHN
        name:
          type: string
          nullable: false
          description: nom de l institution
          example: Muséum National d''Histoire Naturelle
        mandatoryDescription:
          type: string
          format: string
          description: description obligatoire de l institution
          example: institution parisienne
        optionalDescription:
          type: string
          format: string
          description: description optionnel de l institution
          example: institution principale
        logoUrl:
          type: string
          example: https://mediaphoto.mnhn.fr/media/1652113411737lMgX2dOG5396jVeF
        partnerType:
          type: string
          example: DATA_PROVIDER, MEMBER, PARTNER
        url:
          type: string
          format: string
          description: url du site de l'institution
          example: http://mnhn.fr
        specimensCount:
          type: integer
          format: int32
          description: nombre estimé de specimen
      example:
        id: 1
        code: MNHN
        name: Muséum National d'Histoire Naturelle
    InstitutionRequest:
      type: object
      properties:
        code:
          type: string
          format: string
          description: code , libelle court de l institution
          example: MNHN
        name:
          type: string
          format: string
          description: nom de l institution
          example: Muséum National d''Histoire Naturelle
        mandatoryDescription:
          type: string
          format: string
          description: description obligatoire de l institution
          example: institution parisienne
        optionalDescription:
          type: string
          format: string
          description: description optionnel de l institution
          example: institution principale
        partnerType:
          type: string
          format: string
          description: type de partenariat DATA_PROVIDER,PARTNER,MEMBER
          example: PARTNER
        logoUrl:
          type: string
          format: string
          description: url du logo d institution
          example: http://mediaphoto.mnhn.fr/media/1441439763721S1JxP92vYMmLFcfK
        url:
          type: string
          format: string
          description: url du site de l'institution
          example: http://mnhn.fr
        specimensCount:
          type: integer
          format: int32
          description: nombre estimé de specimen
    InstitutionOption:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: id sous forme UUID de l institution
          example: 50f4978a-da62-4fde-8f38-5003bd43ff64
        name:
          type: string
          format: string
          description: nom de l institution
          example: Muséum National d''Histoire Naturelle
...
    InstitutionDashboard:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: id sous forme UUID de l institution
          example: 50f4978a-da62-4fde-8f38-5003bd43ff64
        code:
          type: string
          format: string
          description: code , libelle court de l institution
          example: MNHN
        name:
          type: string
          format: string
          description: nom de l institution
          example: Muséum National d''Histoire Naturelle
        partnerType:
          type: string
          format: string
          description: type de partenariat DATA_PROVIDER,PARTNER,MEMBER
          example: PARTNER
        specimenCount:
          type: string
          format: int
          description: nombre de spécimen dans l'institution
    PartnerResponse:
      type: object
      properties:
        Partner:
          type: array
          items:
            type: string
            format: string
...
    UnexpectedError:
      $ref: '#/components/schemas/Error'
    Forbidden:
      $ref: '#/components/schemas/Error'
    Unauthorized:
      $ref: '#/components/schemas/Error'
    Conflict:
      $ref: '#/components/schemas/Error'
    Error:
      type: object
      description: A common error response
      properties:
        errorCode:
          type: integer
          description: Error Code
        errorMessage:
          type: string
          description: Message
        errorLevel:
          type: string
          description: Level of error
          default: info
          enum:
            - info
            - warning
            - error
            - critical
        errorType:
          type: string
          description: Type of error
          enum:
            - functional
            - technical
        documentationUrl:
          type: string
          description: Url of the documenation
        tips:
          type: string
          description: Tips/help for adressing the error
...